Added code from SVN
Added the code from the old SVN repo, lots of files still need to be
organized so I put them into the "TO BE SORTED" folder.  I figured that
we should at least get the code onto github so that we can all work on
stuff in the same place.