Fixes an error inside standard cache
There was a random error that would occur on different machines that
would cause the tests to fail, but the error itself was random by
nature. However we managed to get the error to be produced on a test
machine and found that it occurred when the pickle was corrupt in just
enough of a way for it not to have an EOF. So we added a capture for
EOFError and that particular issue should occur again.