Adds a changelog to PyPWA
I went ahead and added a changelog to the project using ideas
from http://keepachangelog.com to track changes that might be
useful for users and maintainers over the lifetime of the package.
It's a little more overhead for the developer, but I have a
feeling the package and system maintainers will probably thank us
for it down the line, along with the users if they're into that
sort of thing. It also gives us a solid resource to pull from
once we start putting PyPWA releases back onto our real site.