Adds initial work for the Argument Parser
Moves some of the interfaces out of configurator's options and into
interfaces' plugins to allow the arguments to use a similar interface,
while at adds an interface inside the arguments for Main Options.

Also added a plugin loader and associated tests to load plugins for the
arguments, uses the blank module for testing.