Changes the plugin loaders subclass mechanism
The plugin loader previously would return uninitialized objects to
calling methods. This was fine in most scenarios, but provided little
benefit and allowed for poor programming behaviours.
Now the plugin loader returns initialized objects to the calling
function. All tests have been updated to reflect this and code upgraded
to handle this change.