Flesh out plugin interfaces for Arguments
Updated the Argument Parser plugins to ease adding options
to the provided parser. Plugins add their arguments as a group to the
main parser, while the Main adds its options to the root of the parser;
this is done to give the user an easy way to tell what arguments are
being provided by what feature, and hopefully will allow for the
potential options to be more clear in functionality for the end user.