Update plugin classes
Moves base class declaration from interfaces.plugins back to the
configurator since the argument parser ended up having its own
implimentation.

Adds a more thought out base plugin declaration for the argument parser
that acts like the configurators plugin metadata and setup classes
rolled into one class. This implimentation takes a root parser that is
passed to it and appends its own options to it, but also takes the
namespace and parses it into options into its plugins.
This is done mostly so that the developer can pick the best way to set
the arguments for their plugin without us having to make a complicated
wrapper around Python's argparse.

Also adds a base object for the plugins, this is mostly for static typing,
as this gives us a way to say that we are working with a plugin of some
sort without creating a massive union of the different plugin types.

For example:
type: (Dict[str,BasePlugin) -> None
vs
type: (Dict[str,Union[DataParser,Optimizer,KernelProcessing,..]) ...