Adds full test coverage for masking
Uses the same testing method as the argument parser since it would be
more difficult to interact with the plugin directly while still
maintaining a good test coverage. Also adds various warnings if the
user's masking data is less than or greater than the input file's event
count, along with a collection of pass-fail data to go along with the
sv_test_data.