Moves logging function to initial_logging
Moves the wrapper around the InitialLogger to initial_logging so that
other parts of the program could use, also updated the Start for the
Configurator to use the new logging function. I also removed a lot of
logging code from the processing module and removed some extra
functionality from the InitialLogger.