Modifies test to accept "numpy_unspec_data" file and remove it
The modified test involved testing numpy for changing and writing
to an unspecified file type. In the case of numpy, if a file type
is not given, the file is saved as an numpy array (.npy). This test
accepts a "numpy_unspec_data" and removes a "numpy_unspec_data.npy"
file from memory.