Separates Structured Arrays from Standard Arrays
This is a major change to separate structured arrays from structured
arrays. The reason for this is standard arrays don't contain column data
for the data, and some data types (CSV, EVIL) explicitly require this
information for it to work properly. Now plugins that can handle the
single arrays can get priority for this data and plugins that don't
support this type can be passed for something better.