Changes the way lines are counted
The old way of counting lines was arguably more pythonic, however was a
bottleneck for files greater than 1,000,000 lines, and as such was
causing issues inside the package. The new line search method works by
reading in large chunks of binary data and adding all the newlines found
to that to a single integer and returning that final value.