Refactors loaders and adds bin filtering
This takes the loaders module and turns it into a package called loaders
along with a subpackage called data_loader. It also adds functionality
to the data_loader that will filter events depending on if the value of
that event's bin is zero. This should solve divide by zero issues inside
of the likelihood.