Fixes an issue with the chi-square
This fix was a multi-step process. The multiplier had to be removed from
each likelihood since it was applying the multiplier based on the
Optimizer only, and instead the type of optimizer (Minimizer or
Maximizer) is passed to the likelihood so that the likelihood can
multiply its final value by the multipler instead based both on the
likelihood and the optimizer.