Moves that Data Plugin to a Data Component
The data plugin has been moved to libs, the interfaces to create
another one dismantled, and it's plugins moved to builtin_plugins. The
reasoning behind this is since the data plugin is so complex and it's
interfaces so specific that there was little reason for a user to
provide their own since the data plugin supports plugins itself.