Replaces os with Path from pathlib
Pathlib is the new Python library for handling file paths. Instead
operating on a string that represents a path like os and io do, pathlib
instead treats each path like an object. This grants us some code that
is more straight forward and easy to understand.
To see more information about pathlib, see here:
https://docs.python.org/3/library/pathlib.html

Since the module name for Pathlib is different for python versions
before 3.4, we import pathlib.Path and Pathlib.PurePath on the
package level to ensure that we have access to the module everywhere
in the package. It's also added to the setup so that pathlib2 can be
installed on systems that need it.