Add Vector and Particle support
This adds vectors and particles to PyPWA. The vectors are reworked from
the vectors that were originally written by Stephanie Bramlett for the
ISOBAR.

The new vectors utilize Numpy arrays to accelerate calculation over
each event. Each vector is a structured Numpy array with each column
being a component of that vector: 3 Vectors having 'x', 'y' and 'z'
columns, with 4 vectors adding a 4th 'e' column. Each row in the vector
represents an event for that vector. Included specific features are:

- Add, subtract, multiply, dot, vectors together or by scalars
- Iterate over the events in a vector using 'for event in vector:'
- Request components of the vector directly using Vector['x'], ...
- Ability to return a complete copy of the Vector, instead of the
  reference
- Ability to get the array from the vector
- Calculate length, length squared, theta, phi, cos theta, and sin theta
- Access each component through a property, Vector.x, Vector.y, ...
- 4 vector can return a 3 vector of it's x, y, and z values

The new Particle class extends from the 4 vector class and adds
additional properties to track the Particles GEANT Id and the particles
name. Features that the Particle class have right now are:

- Get specific event by index using Particle[index]
- Get the number of events with len(Particle)
- Stores the charge, name, and ID of particle
- Get X, Y, Z, and E components with Particle['x'], Particle['y'], ...
- Clean(ish) representation of the particle with repr(Particle)
- Split Particles into chunks using Particle.split

Lastly the new ParticlePool tracks all particles that were loaded in
from a single file, allows the user to request specific particles from
the pool, and iterate over both particles in the pool and events.
Specific features are:

- Iteration over events using 'for event in ParticlePool:' or
  ParticlePool.iterate_over_events()
- Iteratation over particles using ParticlePool.iterate_over_particles
- Return number of particles in the pool using either len(ParticlePool)
  or ParticlePool.particle_count
- Return number of events stored in the pool with
  ParticlePool.event_count

This also updates the components to support the new data type. Data
Processor is reworked to support vector data instead of the old
depreaciated stacked array format, and GAMP reader, writer, and memory
have been re-written for the new Vector format; and Process adds a new
_data_split class to handle ParticlePool. Finally, PyMask has had it's
tests updated ensure compatibility with ParticlePool data.