Use boolean flags instead of data in Data's Writer
Up to now, since we didn't have a clear way to define particle data, we
used the shape of the Numpy array to determine the best writer to select
when writing out data. This worked in most situations since the majority
of the data worked in PyPWA was parsed or dumped, meaning that with the
expectation of the Masking utility, data was already in memory before it
needed to be written.
With the future addition of the binner, this assumption in no longer
valid. The binner requires several writers to be opened and only one
reader, potentially not even in the same process. This means that we can
no longer assume that we will have an example event to select a writer
from, we will need to figure out the data type ahead of time, then open
all the writers from there.
This is plausible now. The readers and writers will now record whether
they are reading or writing particle data. With the writer, when
requesting a new writer you no longer need to provide data in order for
it to work, instead you tell it either if the data is particle data, or
if the data is a standard Numpy array without columns. Standard columned
data is always assumed true unless something else is provided.