Fix issue where GAMP event count was incorrect
The GAMP iterator would return the wrong event count when called, it
would take the number of lines divided by the number of particles, not
the number of lines divided by the number of lines per event. This would
return incorrect values because the number of lines is always one
greater than the number of particles.