Improve configurator
The configurator gained the ability to have dictionaries packed into a
list as long as the dictionary template is the same for each dictionary
inside the list. There is also been some altering to the configurator to
allow it to the skip writing the python funtion file if there will be
nothing in the file.