Add pybinning and pytable support
This is a massive change for PyPWA. There is still a lot of work that
needs to be done, but succently this is the first step in a massive
refactoring of PyPWA 2 so that it may support GUIs and the ISOBAR model.
Unfortuantely this means it's no longer compatible with PyPWA 2, so this
marks the first in-tree alpha of PyPWA 3.

Signed-off-by: Mark Jones <maj@jlab.org>