Combine cache package into single module
This is a massive refactoring of the cache package. Now all the
functionality of the cache package is contained in a single short
module, with each of its classes refactored to be easier to read and
interact with. Now also contains documentation to help new users figure
out how to use it.