Refactor Data plug-ins and process package
This refactoring furthers the attempts to simplify some of the code to
be a little easier to use. In doing so, this removes support for
precision selection as this is an unused feature. This also combines
file_processor and _plugin_finder into a single file, and combines the
metadata files with their corresponding process file.