Combine process modules into single package
This is a minor refactoring of the process package. This combines all
the files into a single file, breaks up some objects into a collection
of functions, and combines the duplex and simplex process into a single
Process object.

This also updates the documentation and the tests for the process
module.