Update data management for pypwa
This is a massive update that improves greatly on the data libraries for PyPWA
by expanding on the usability of tables through ProjectDatabase and passing
through helpful functions and objects to PyPWA's init.
As a positive efect, many of the progs modules have refactored to be cleaner
and easier transparent.

Things that are new:
 - PyPWA now imports all useful functions for the user, so that when the user
    imports the package inside IPython or Jupyter, they'll immediately have
    access to all the tools they need to get started.
 - Reader and Writer now share the path of the file they're operating on.
 - Binning now works by both fixed variable count, and widths.
 - Package info is now contained in info.py.

Things that changed:
 - SlotTable has been changed to ProjectDatabase. The new package now makes
    more assumptions about the users data and as a result should be more
    powerful and easier to work with.
 - Documentation has been completely reworked to a more straightforward layout.