Add support for fitting, pandas, and more
This is a large overhaul of PyPWA to support Pandas all the way to the
core, as well as to add a handful of features and fixes that have been
numerously requested.

This patch adds support for:
- Fitting, using the new Likelihood objects with minuit, others coming
  in the future.
- Refactored Vectors with reactions rolled into ParticlePool
- Pandas added everywhere, file objects load pandas instead of
  sturctured arrays, and all libraries now support both pandas
  and structed arrays
- Initial support for resonance. Consider this pre-alpha until more
  feedback is collected.
- Binning in memory is now supported, and more options will be added in
  the future.
- PyPWA can now be used from a script or from Jupyter notebooks much
  easier, just import and start using it like you would numpy, scipy, or
  any other package you may already be familiar with.

The major changes are a switch to DataFrames and brand new Likelihood
objects. The DataFrames will work similar to structured arrays, but
different enough that you can't assume it's a perfect drop in
replacement. Be sure to read the pandas documentation on anything that
isn't immediately clear.

The other change is the new Likelihoods. These allow you to take your
amplitude, package it inside one of our already predefined likelihoods,
and use it with any optimizer you like while still being able to take
advantage of your systems multiprocessing capabilities.