Patch speed issue with kv parser
The KV Parser was taking a pd.Series and placing it directly into a
pregenerated pd.DataFrame, then returning the result. This is
significantly slower than the direct numpy route.

The fix was to allow the KV Reader to return just the array and to place
that array directly into another numpy array, and then returning a
DataFrame constructed from that resulting numpy array.