Patch multiple processing and data issues
This patch fixes processing erroring, gamp parsing speed, cache loading,
cache overrides, binning, likelihoods, and pipes.

The processing module needed a few design tweaks to properly handle
pipes in a way that wouldn't result in a OSError. Now, the stop method
has been removed from the ProcessInterface, and has been replaced with
a close method that properly shuts down the processes, as well as close
the effected pipes. SmartProcess has been updated as well, so that when
the process closes, or the process encounters an exception, the pipe for
that process will be closed as well. When close is properly called on
ProcessInterface after the objects are done being used, it shouldn't be
possible to encounter the OSError anymore.

Processing also supported sending a ProcessCode when an error was
encountered, but the interfaces wouldn't search for the ProcessCode, and
would instead throw an error when it encountered it, suppressing the real
error. Now, the interfaces correctly search for an handle the error, as
well as forwarding the exception so that the user can patch their
amplitude.

Caching has been patched as well. You can now forcefully disable the
cache, or overwrite the cache from read and write. The issue with the
caching module resulted from the cache package would overwrite the
created cache package when it was loaded, which meant that when hashes
were compared, they'd always result in the hashes appearing to match.

Gamp memory parser was updated to allow for quicker parsing of Gamp
files. DataFrames are used internally to store the values, but changing
a DataFrames values after it's been created is a relatively time
consuming process. It's significantly faster to create a numpy array and
update its values with what is in the file, and then convert that numpy
array into a DataFrame. This is effectly what Gamp does now, and is
similar to the method used by the Kv parser.

Binning had an issue where it return N bins - 1, now, it will return the
correct number of bins.

The Extended Log Likelihood would return the wrong result due to a
miscalculation in the likelihood. This has been corrected.