Patch cache package and likelihoods
When a cache package isn't loaded, either due to a broken pickle, or if
the cache doesn't exist, it wouldn't set the loaded cache package at
all, so it would fail when it attempted to make a comparison.

The likehood objects don't extend the ProcessInterface objects, instead
they wrap them, so they had to be updated to use the new close
mechanism, otherwise, they would fail.