Fix cache and fit errors, improve binning
This attempts to fix the ongoing caching issue, speed up processing,
improve likelihoods, and patch minuit.

To fix the caching issue, many methods were simplified to try to improve
the logic of the caching module, and the module now checks to ensure
that there was data loaded from the package before it returns it. Before
the validation check would only ensure that the package was loaded and
that the file hashes matched, now it goes as far to check that there was
data loaded from the cache as well. This should prevent NoneType errors
in the future.

In order to fix the issue the OSError with the processing module, the
ProcessInterface object now closed the Pipes as well as the Processes.
This should allow for significantly more processes to be spawned without
an issue. The mechanism to check that the processes have shutdown savely
before terminating them should be improved now, so there will hopefully
be less time waiting between shutting a group of processes and opening
another group in the case of doing binned runs.

Minuit used to close the processes when it was done, this is no longer
the case. The option has been removed and the user will be expected to
close the likelihoods themselves.

Likelihoods have been greatly improved. Instead of data being a
parameter that accepts a dictionary with the neccassary data values,
each expected or optional data value. This should make it a little more
clear for the user to select which values they want.
Likelihoods have also now support being initialized using with, you can
now:
```
with Likelihood(...) as likelihood:
    return minimize(likelihood)
```

Lastly, binning has been improved to support trimming off data by
setting lower and upper bounds.