Add support for binning by fixed widths in memory
Binning by fixed with works by creating an index array of the binning
data, and then using that to select values from to place into each bin.
This should mean that each bin should have the same length, and values
should be sorted correctly through the bins without disturbing the order
of the events.