Add sample selection to bin_by_range
You can now have a randomly selected number of events for each bin with
bin_by_range by specifying sample_size in the functions parameters.
This should allow users to force fixed number of events for each bin
while still being a valid sample.

Both bin_by_range and bin_with_fixed_width now support passing a column
name directly to the function as the bin_series instead of a seperate
array. So if you are binning your data by a column in your datafame,
you are now able to pass the name of the column instead of the array
extracted from the dataframe.