Add support for masking all Vector classes
Vector's can now be masked like DataFrames and Numpy's Arrays. This
includes ThreeVectors, FourVectors, Particles, and the ParticlePool.
ParticlePool's behavior is different from traditional data structures,
the length will return the number of particles it's containing, and when
passed an index it will return a particle, not an event; but can accept
an event mask that will mask each particle, and return a new particle
pool with each particle masked with that event mask.