Update documentation
This is the beginning of adjusting the documentation to be better for
the PyPWA users. We'll be adding more documentation throughout the
summer as user's have confusion about how to use a certian function, or
to complete a specific function.

This also includes updated GitHub documents, so now we'll have templates
for issues, a contributors file, and an up to date readme.