Replace pandas with numpy inside Vectors
Using Pandas inside the vectors seemed like a good idea because of the
benefits of Pandas. However, Pandas ended up being to detrimental to the
speed of the program. This partially reverses the decision to swap to
pandas. You can still access the data as a DataFrame, but the values are
now stored internally as numpy arrays