Update documentation
ReadTheDocs wouldn't work with our Sphinx setup because it doesn't
install PyPWA before it builds the documentation. The solution was just
to hardcode the values into the Sphinx documentation.

This also patches the readme to exclude the Beta text for Summer 2020,
and to include the grant numbers from NSF that pays for our reseachers.

Cleans up the CHANGELOG and adds formatting for libraries.