Release PyPWA 3.1.0
This is a big overhaul to PyPWA's documentation. PyPWA now has proper
documentation that is listed on ReadTheDocs instead of just inline
documentation and the short how to install documentation that was
originally found on the ReadTheDocs. This is just the first major
batch of documentation, and more will come with time as we collect
feedback.

This release also contains an update to the data module that allows
users to select if they want pandas or numpy arrays as their return
when loading and working with data. The memory plugins assume the
return will be in Numpy arrays, and the memory wrapper will wrap the
returns in DataFrames or Series if the user sets use_pandas to true.
The readers must impliment handling the Pandas returns on their own.

Some helping functions have been added as well that should aid moving
data to contiguous containers for use with Cython and C/Fortran
functions. There is also a new experiemental file format to handle
particle data that should eventually replace the GAMP file format.

As general upkeep changes:
- Minuit's wrapper has been updated for the new API
- Vectors havee been swapped from dataframes to contiguous arrays
- ReadTheDocs requirements have been updated to latest requirements