Fixes #124 incorrect value order and broken math
This adds tests for the Vectors to catch and value sorting errors.
Originally the order of the variables was x, y, z, e, and eventually
that order was changed to e, x, y, z, since that is the standard
ordering of values. There was errors produced from that change, and
this should patch those errors.

Next, the sanitization function was well tested with arrays, but was
loosely tested for single values since in our use-cases we did not use
any other than inputs. This should patch those issues out as well,
and there is now more testing to catch potential errors that might
arrise in the future from that.

Lastly, I added iPython and Jupyter pretty output while I was here,
so with using iPython or Jupyter there should be some nice output
to explain what the Vector stores. If you want to see the raw data
inside the container, you can still use repr(vector).