Patches for GAMP parsing and Particle display
The memory module of GAMP would use a dictionary to store the particle
data as an intermediate step, using the particle IDs as the key in the
dictionary. This worked great, until there were multiple particles with
the same ID. This patches that out by using the index of the particle
instead. This assumes that each event will have particles appear in the
same order consistently throughout the file, which _should_ be a safe
bet.

There was also an issue with some rushed out code involving the how
Particles are visually represented in iPython and Jupyter. The
underlying data structures were safe, but when the user would try to
view the values, errors would occur. This should also patch those
issues. However, there are no improvements in Testing with these sort of
functions since at this time I have not found a reliable way to test
that the display function is operating correctly.