Correct masking with particles
There was an issue where the Particle class would not pass the
masking/indexing array through to the Numpy Arrays when the object was
being masked/indexed. This meant that it'd basically return a copy of
the Particle with no transform on it. This patch solves that by putting
the mask back into the correct place.