Solve Numpy and Python depreciation warnings
For numpy, it was more of the float issue that needed to be resolved. A
quick switch from numpy.float to numpy.float64 guarenteed the same
behaviour as before.

For Python, it seems that how the interpretor interprets strings for
docstrings has changed, and was reading the latex switches as escape
characters. So I escaped the escape characters and now those
depreciation warnings have been resolved as well.