Fix regression in Gamp writing
There was an issue with GAMP not writing data again that's a regression
from the patches for v3.2.0. If I had to guess, the patches to avoid
deprecations from Numpy changed the internal data types for the vectors
in a minor way that caused the writer to break. In v3.2.0 it's assumed
the internal values when extracted by a single value would be a scalar,
the patch to 3.2.1 assumes it's an array with a single element, and then
one of the changes in 3.2.2 changed that internal value back to a
scalar.

This patch doesn't revert to assuming the data is a scalar, instead we
now wrap the writing data with `float` so that regardless of whether
it's a scalar or an array with a single element. Hopefully that means
this particular issue will no longer present itself.

As a bonus, ParticlePools can now be compared against each other. You
can now `p1_pool == p2_pool` and recieve a boolean result depending on
if they are equal or not. While there is not a lot of praticle use for
this for Physics outside of sanity checking, it does make the process of
unit testing the libraries much more manageable.