Improve pytest automation
This does a few things for us, pytest can now watch both the tests/ and
PyPWA/ directories for changes, and automatically rerun tests on
changes.

This also sets the testing directory, and defaults for coverage.
Eventually there should be multiprocessing for the individual tests as
well, as soon as overlapped resources are solved.