Add Anaconda Environments
This is the beginning of the Anaconda environments around PyPWA. There
are two to start- a stable environment called `anaconda-environment.yml`
and an unstable environment called `dev-environment.yml`.

The stable environment should be used by anyone using the package, the
unstable environment should be used by anyone developing on PyPWA
directly.