Move cache to same directory as data
Before PyPWA would bury the cache file into the cache folder inside
the users home directory, but now the cache file will be placed
directly inside the same directory as the data. This should allow it to
be more obvious to the user what data is being cached, or not cached,
and should allow users to more easily view what is happenning inside the
cache file, if they are curious.