Add optional dependency code to EMCEE
This adds some optional dependency with EMCEE to the MCMC code, this
should allow PyPWA to function as expected even when EMCEE is not
installed, but will inclde the MCMC functionality when the dependency is
installed.

If you use the anaconda environment included with the package, MCMC will
be enabled by default, however if you install the package by hand into
Anaconda, EMCEE will not be included since it is not inside the base
repos.

Lastly, I reformatted the code to more closely follow the PEP8 format.
Hope you don't mind Peter.