Update CHANGELOG to hold new features
A few changes have been added over the past few days, namely an updated
documentation environment, CuPy swapped with PyTorch, bug chasing from
RTD, and finally support for multi-GPU configurations. There is still
more work that needs to be done, but this might be officially in a state
where we can start collecting feedback on the package.