Downgrade Python 3.10 to 3.9
Unfortunately Anaconda currently has not moved all current packages to
Python 3.10, maybe for stability issues or otherwise. This means that as
of now, it is not possible to produce an environment that contains all
the packages used by the physicists that use our software. By
downgrading to Python to 3.9, all the packages can resolve through conda
without resorting to third-party channels.