Update Minuit and Likelihood code
This should solve the issue around iMinuit once and for all- Now, the
optimizer is returned without limits or fixed values being set, and it's
up to the user to set those values, and then call migrad. To make things
a little easier for the user; a new `TYPE` parameter has been added to
each of the likelihoods, if the Likelihood is a LogLikelihood, the TYPE
is LogLikelihood, if it's ChiSquared it's ChiSquared, and lastly it's
Other for none of the above. This should cover everything, but it's an
Enum so it should be easy to expand the list in the future.