Fixed issue with bin by range samples
The sample function was supposed to not only take a sample of bins for
each bin, it was supposed to rearrange the bins to be in a random but
consistent with itself order. However, there was a bug introduced where
it would only remove a single random bin.

This fixes that issue by appropriately rearranging the events as
expected, without further masking the resulting arrays.