Add multi-GPU support with threads
If this functions as intended, this will be the final big feature added
to PyPWA for a while. This introduces the ability for code to run on
several GPUs concurrently using Threads. Each Thread will talk to its
own GPU, and if enough of the Thread's time is GPU bound, then this
should yield some fairly performant code.