Clean up Flags for Amplitudes, expand Torch
This cleans up the flags so it's more clear what state the Likelihood
will end up in after it builds itself than it was before. The big,
final, piece of this was adding a `DEBUG` flag that is effectively the
same as setting `USE_MP` to false before, but now more explicitly. You
can still reach a `DEBUG` state through setting other flags to False,
but it's much faster to just use the `DEBUG` flag. It's more
consistent too, since setting `USE_THREADS` and `USE_MP` to False will
effectively be the same as `DEBUG` until you set `USE_GPU` to True, then
you'll end up in processes again.

This just makes the desired functionality more explicity, plus it's
better from a user's prospective to set `DEBUG` to True when debugging,
than a handful of other flags to False.

I also went ahead and split out Torch as it's own method on each of the
Likelihoods to prevent Python's chaotic branching on each likelihood
call. While this should make the likelihoods run faster, it'll mostly
benefit those reading this code later, since it should be more clear
what is happening where. Assuming they take the time to read the setup
function.

Lastly, Torch and GPU support are now divorced thoroughly, since Torch
can be used by iMacs to speed up execution, and iMacs don't have CUDA
GPUs. Now, to get to the GPU you must set `USE_GPU` to True. This should
be pretty clear to even the densest of users.