Use Queues over Pipes with Threading
This is more-or-less a nice to have feature, than a neccassary feature.
Pipes _can_ function with Threads, Pipes allow for communication of
information from one point to another, just as Queues do. However,
Queues do not work with Multiprocessing, on Pipes do, and Pipes can not
move any data that can not be Pickled.

Queues can move Pickled data.

This means that if someone is using data or an optimizer that does not
work with Pipes, because the data can't be pickled; when using Threads,
the data will be packaged in Queues instead, which do not have the
caveats that Multiprocessing Pipes share.