Change THREAD default to 0
Thread was originally -1, inspired by Torch and how it determines
the number of GPUs available. -1 for none, 0 for 1, etc... however,
this does not make sense in execution, since it's counting the
number of threads, not GPUs. It's possible for Threads to be much
greater than the number of GPUs available. So it's now set to 0,
since at the very least, the Amplitude can be running in the master
thread.