Patch Data Module to read cache objects
This has been a bug in PyPWA for much of PyPWA 3, or at least
since intermediate caching was introduced. The now old intermediate
cache flags replaced the old enable_cache flag, since the duty of
enabling or disabling the cache was moved out of the cache object's
responsibility. What the slight flag mixup resulted in was the
Parser would tell the Cache object to treat the input file as an
intermediate cache instead of a cache object; and since there
didn't exist an intermediate cache it would always fail on
FileNotFound, however the write function of the Cache did the
opposite of the Read module and assumed it assumed intermediate
to be False. This would result in behavior where the cache would
constantly fail to load, but always successfully write.

This, after what as likely been a great deal of time, has been
remedied. The cache object now always assumes Intermediate to be
true, and the Loader and Dumper explicitly set Intermediate to
False.

This should mean that anyone using the Cache module as an
intermediate cache will only need to provide the path, and
everywhere else can just hard-encode bypassing that behaviour.